feat(api): a Pick'em week is one immutable submission
The write path stops being an editor. A member submits the full required set
for a week once, and after it lands no pick in that week can be changed,
replaced, or removed (ADR-0018 decision 1). `submitPickemPicks` only ever
inserts: the replaceable/retained computation, the `picksAllowed - retainedCount`
cap arithmetic, and the delete-then-reinsert re-pricing all describe an
operation that no longer exists. `repickPickemPick` and `POST .../repick` go
with them — PKM-7's substitute endpoint was rule 1's inverse, and an inverse of
a deleted rule has nothing to invert — and so do the refusals only editing could
produce, `pick_not_found` and `pick_not_replaceable`, from `PICKEM_REFUSAL`,
both handler maps, and `ERROR_CODE`.

The required set is sized against the *unlocked* pickable slate, evaluated
inside the submitting transaction (ADR-0018 decision 2). The naive reading —
`submissions.length === picksAllowed` — would be harmful: `pickable` does not
exclude a game that has already kicked off, so once the week's first game
started, a member who had not yet submitted would be asked for more picks than
the slate can supply and would be locked out of the week permanently. That is
an implicit league-wide weekly deadline, which the ADR explicitly refuses.
Games that locked before the member submitted are forgone; they were never
picks, so nothing scores.

That rule lives in one place, `requiredPickemPickCount` in `packages/schemas`,
because SIMP-10's Save button must gate on the same number — the same reason
`pickemSettingsInvalidatePicks` is a shared pure function. The API passes
`settings.picksPerWeek` and the web will pass the wire's `picksAllowed`; its
docblock proves those agree, since `min(min(p, pickable), unlockedPickable)`
and `min(p, unlockedPickable)` are the same number. It computes a domain
answer, so it is unit-tested. The wire field `picksAllowed` keeps exactly the
meaning it had.

Refusal order inside the transaction is deliberate: `already_submitted` (409, a
conflict with state the member cannot change) is checked before any per-game
validation, so a second attempt gets the honest reason rather than a stale-spread
complaint about a sheet that was never going to be accepted. Then the size
check — `too_many_picks` (400) over, `pick_set_incomplete` (400) under, exact
mirrors. Every per-game check is unchanged: `game_not_in_week`, `pick_locked`,
`game_not_pickable`, and ATS's `spread_unavailable` / `spread_stale`. The line
still moves between page load and submit, so the first submission's ATS
handshake survives intact; immutability removed the second submission, not that.

`lockLeagueMemberRow` stays and is more load-bearing than before, not less: it
is now what makes "one submission" true, since two requests racing on an empty
week would otherwise both read no picks and both insert. The duplicate-pick
unique violation stays as the DB backstop.

The settings-invalidation reset (ADR-0015 rule 3) survives and gains a second
job: it is now the only path by which a member re-submits a week, because it
deletes the very picks `already_submitted` keys on — and therefore the only
remedy for a submission a member regrets. Deliberately narrow: commissioner-only,
pre-start-only, refused with `picks_locked` once anything has locked.

Tests follow the rules. Gone: wholesale-replace, empty-array-clears, the
retention boundary, and cap-remaining-slots — those rules no longer exist.
Added, each named: a second submit refused with the first set intact; over- and
under-sized sets; an empty submission; a member submitting after the week's
first kickoff succeeding with a full set of what is still unlocked (the ADR's
decision 2, and the most important case here); and a raised Picks Per Week
clearing a submitted week so the member can submit again end to end. The lock,
spread-stale, spread-unavailable, duplicate, week-range, rollback and
league-concluded cases stay, retargeted to full-size sheets so a refusal can't
be confused with a size complaint. `standings-repick.test.ts` loses its repick
half and becomes `pickem-standings.test.ts`.

`apps/api` typecheck still carries SIMP-9's `GAME_STATUS.MOVED` errors, and
`apps/web/src/api/pickem.ts` now also references the two retired error codes;
both clear in the next commit and in SIMP-10.

Refs SIMP-8
