+++ apps/api/test/pickem-picks.test.ts
@@ -429,6 +429,9 @@ describe("GET /api/leagues/:leagueId/pickem/weeks/:weekId/picks", () => {
     const memberB = await createAuthenticatedUser(auth, { username: "member_b" });
     const league = await insertLeague(db, {
       seasonId,
+      // Two picks a week against a three-game slate, so each member's one
+      // submission can be a different pair rather than the whole slate.
+      settings: { ...DEFAULT_PICKEM_SETTINGS, picksPerWeek: 2 },
       members: [
         { userId: memberA.user.id, role: MEMBER_ROLE.COMMISSIONER },
         { userId: memberB.user.id, role: MEMBER_ROLE.MEMBER },
@@ -614,7 +617,158 @@ describe("PUT /api/leagues/:leagueId/pickem/weeks/:weekId/picks", () => {
     expect(await res.json()).toMatchObject({ error: "league_not_found" });
   });
 
-  it("saves a happy-path submission, and a later different set replaces it wholesale", async () => {
+  it("saves the week's full set", async () => {
+    const { league, weekIds, gameIds, memberA } = await seedPickemLeague();
+    const weekId = weekIds.get("regular:1")!;
+    const [g1, g2, g3] = gameIds.get("regular:1")!;
+
+    const res = await putPicks(memberA.cookie, league.id, weekId, {
+      picks: [
+        { gameId: g1, side: PICKEM_PICK_SIDE.HOME, spread: null },
+        { gameId: g2, side: PICKEM_PICK_SIDE.AWAY, spread: null },
+        { gameId: g3, side: PICKEM_PICK_SIDE.HOME, spread: null },
+      ],
+    });
+    expect(res.status).toBe(200);
+    const own = ((await res.json()) as PickemWeekPicksResponse).members.find(
+      (m) => m.userId === memberA.user.id,
+    )!;
+    expect(own.picks.map((p) => p.gameId).sort()).toEqual([g1, g2, g3].sort());
+    expect(own.picks.find((p) => p.gameId === g2)?.side).toBe(PICKEM_PICK_SIDE.AWAY);
+  });
+
+  it("409s already_submitted on a second submission for the same week, leaving the first untouched", async () => {
+    const { league, weekIds, gameIds, memberA } = await seedPickemLeague();
+    const weekId = weekIds.get("regular:1")!;
+    const [g1, g2, g3] = gameIds.get("regular:1")!;
+    const fullSet = [g1, g2, g3].map((gameId) => ({
+      gameId: gameId!,
+      side: PICKEM_PICK_SIDE.HOME,
+      spread: null,
+    }));
+
+    expect((await putPicks(memberA.cookie, league.id, weekId, { picks: fullSet })).status).toBe(
+      200,
+    );
+
+    // A week is one submission (ADR-0018) — even an identically-sized, legal
+    // set is refused, and the sides the member first chose survive.
+    const second = await putPicks(memberA.cookie, league.id, weekId, {
+      picks: fullSet.map((pick) => ({ ...pick, side: PICKEM_PICK_SIDE.AWAY })),
+    });
+    expect(second.status).toBe(409);
+    expect(await second.json()).toMatchObject({ error: "already_submitted" });
+
+    const after = await getPicks(memberA.cookie, league.id, weekId);
+    const own = ((await after.json()) as PickemWeekPicksResponse).members.find(
+      (m) => m.userId === memberA.user.id,
+    )!;
+    expect(own.picks).toHaveLength(3);
+    expect(own.picks.every((p) => p.side === PICKEM_PICK_SIDE.HOME)).toBe(true);
+  });
+
+  it("400s pick_set_incomplete when the set is smaller than the week requires", async () => {
+    const { league, weekIds, gameIds, memberA } = await seedPickemLeague();
+    const weekId = weekIds.get("regular:1")!;
+    const [g1, g2] = gameIds.get("regular:1")!;
+
+    // Three pickable, unlocked games and a cap of 5 — two picks is a partial
+    // sheet, and a partial sheet can't be topped up later.
+    const res = await putPicks(memberA.cookie, league.id, weekId, {
+      picks: [
+        { gameId: g1, side: PICKEM_PICK_SIDE.HOME, spread: null },
+        { gameId: g2, side: PICKEM_PICK_SIDE.HOME, spread: null },
+      ],
+    });
+    expect(res.status).toBe(400);
+    expect(await res.json()).toMatchObject({ error: "pick_set_incomplete" });
+    expect(await pickCountFor(league.id)).toBe(0);
+  });
+
+  it("400s pick_set_incomplete on an empty submission — there is no way to submit nothing", async () => {
+    const { league, weekIds, memberA } = await seedPickemLeague();
+
+    const res = await putPicks(memberA.cookie, league.id, weekIds.get("regular:1")!, { picks: [] });
+    expect(res.status).toBe(400);
+    expect(await res.json()).toMatchObject({ error: "pick_set_incomplete" });
+  });
+
+  it("lets a member who arrives after the week's first kickoff submit what's still unlocked (ADR-0018)", async () => {
+    const { league, weekIds, gameIds, memberA } = await seedPickemLeague();
+    const weekId = weekIds.get("regular:1")!;
+    const [, g2, g3] = gameIds.get("regular:1")!;
+
+    // The week's first game has kicked off; g2 and g3 have not. The required
+    // set is what can
+    // still be picked — sizing it against picksAllowed (3, since a kicked-off
+    // game is still `pickable`) would lock this member out of the week for
+    // good, which is the implicit weekly deadline the ADR refuses.
+    const res = await putPicks(
+      memberA.cookie,
+      league.id,
+      weekId,
+      {
+        picks: [
+          { gameId: g2, side: PICKEM_PICK_SIDE.HOME, spread: null },
+          { gameId: g3, side: PICKEM_PICK_SIDE.AWAY, spread: null },
+        ],
+      },
+      appAfterKickoff,
+    );
+    expect(res.status).toBe(200);
+    const own = ((await res.json()) as PickemWeekPicksResponse).members.find(
+      (m) => m.userId === memberA.user.id,
+    )!;
+    // g1 is forgone, not owed: it was never a pick, so it scores nothing.
+    expect(own.picks.map((p) => p.gameId).sort()).toEqual([g2, g3].sort());
+  });
+
+  it("409s pick_locked when a submitted game has already kicked off", async () => {
+    const { league, weekIds, gameIds, memberA } = await seedPickemLeague();
+    const weekId = weekIds.get("regular:1")!;
+    const [g1, g2] = gameIds.get("regular:1")!;
+
+    // A sheet assembled before g1 kicked off: the right size for the slate the
+    // member saw, but naming a game that has since locked.
+    const res = await putPicks(
+      memberA.cookie,
+      league.id,
+      weekId,
+      {
+        picks: [
+          { gameId: g1, side: PICKEM_PICK_SIDE.HOME, spread: null },
+          { gameId: g2, side: PICKEM_PICK_SIDE.HOME, spread: null },
+        ],
+      },
+      appAfterKickoff,
+    );
+    expect(res.status).toBe(409);
+    expect(await res.json()).toMatchObject({ error: "pick_locked" });
+    expect(await pickCountFor(league.id)).toBe(0);
+  });
+
+  it("409s pick_locked exactly at the kickoff instant — the boundary is locked", async () => {
+    const { league, weekIds, gameIds, memberA } = await seedPickemLeague();
+    const weekId = weekIds.get("regular:1")!;
+    const [g1, g2] = gameIds.get("regular:1")!;
+
+    const res = await putPicks(
+      memberA.cookie,
+      league.id,
+      weekId,
+      {
+        picks: [
+          { gameId: g1, side: PICKEM_PICK_SIDE.HOME, spread: null },
+          { gameId: g2, side: PICKEM_PICK_SIDE.HOME, spread: null },
+        ],
+      },
+      appAtKickoff,
+    );
+    expect(res.status).toBe(409);
+    expect(await res.json()).toMatchObject({ error: "pick_locked" });
+  });
+
+  it("re-opens the week when a settings change clears the member's picks, and accepts a fresh full set", async () => {
     const { league, weekIds, gameIds, memberA } = await seedPickemLeague();
     const weekId = weekIds.get("regular:1")!;
     const [g1, g2, g3] = gameIds.get("regular:1")!;
@@ -622,396 +776,80 @@ describe("PUT /api/leagues/:leagueId/pickem/weeks/:weekId/picks", () => {
     const first = await putPicks(memberA.cookie, league.id, weekId, {
       picks: [
         { gameId: g1, side: PICKEM_PICK_SIDE.HOME, spread: null },
-        { gameId: g2, side: PICKEM_PICK_SIDE.AWAY, spread: null },
+        { gameId: g2, side: PICKEM_PICK_SIDE.HOME, spread: null },
         { gameId: g3, side: PICKEM_PICK_SIDE.HOME, spread: null },
       ],
     });
     expect(first.status).toBe(200);
-    const firstBody = (await first.json()) as PickemWeekPicksResponse;
-    const firstOwn = firstBody.members.find((m) => m.userId === memberA.user.id)!;
-    expect(firstOwn.picks.map((p) => p.gameId).sort()).toEqual([g1, g2, g3].sort());
 
-    // Re-submit a wholly different set — the old picks must be gone, not merged.
-    const second = await putPicks(memberA.cookie, league.id, weekId, {
-      picks: [{ gameId: g2, side: PICKEM_PICK_SIDE.HOME, spread: null }],
+    // Raising Picks Per Week strands a submitted member under submit-once —
+    // they can never reach the new size — so the reset clears their picks
+    // (ADR-0018 decision 1 / ADR-0015 rule 3). That reset is the only path
+    // back into a week the member has already submitted.
+    const patched = await patchLeague(memberA.cookie, league.id, {
+      settings: { ...DEFAULT_PICKEM_SETTINGS, picksPerWeek: 8 },
     });
-    expect(second.status).toBe(200);
-    const secondBody = (await second.json()) as PickemWeekPicksResponse;
-    const secondOwn = secondBody.members.find((m) => m.userId === memberA.user.id)!;
-    expect(secondOwn.picks).toHaveLength(1);
-    expect(secondOwn.picks[0]).toMatchObject({ gameId: g2, side: PICKEM_PICK_SIDE.HOME });
-  });
+    expect(patched.status).toBe(200);
+    expect(await pickCountFor(league.id)).toBe(0);
 
-  it("clears the member's unstarted picks when submitting an empty array", async () => {
-    const { league, weekIds, gameIds, memberA } = await seedPickemLeague();
-    const weekId = weekIds.get("regular:1")!;
-    const [g1] = gameIds.get("regular:1")!;
-
-    await putPicks(memberA.cookie, league.id, weekId, {
-      picks: [{ gameId: g1, side: PICKEM_PICK_SIDE.HOME, spread: null }],
+    const resubmit = await putPicks(memberA.cookie, league.id, weekId, {
+      picks: [
+        { gameId: g1, side: PICKEM_PICK_SIDE.AWAY, spread: null },
+        { gameId: g2, side: PICKEM_PICK_SIDE.AWAY, spread: null },
+        { gameId: g3, side: PICKEM_PICK_SIDE.AWAY, spread: null },
+      ],
     });
-    const res = await putPicks(memberA.cookie, league.id, weekId, { picks: [] });
-    expect(res.status).toBe(200);
-    const body = (await res.json()) as PickemWeekPicksResponse;
-    expect(body.members.find((m) => m.userId === memberA.user.id)?.picks).toEqual([]);
-  });
-
-  it("409s pick_locked when a submitted game has already kicked off", async () => {
-    const { league, weekIds, gameIds, memberA } = await seedPickemLeague();
-    const weekId = weekIds.get("regular:1")!;
-    const [g1] = gameIds.get("regular:1")!;
-
-    const res = await putPicks(
-      memberA.cookie,
-      league.id,
-      weekId,
-      { picks: [{ gameId: g1, side: PICKEM_PICK_SIDE.HOME, spread: null }] },
-      appAfterKickoff,
-    );
-    expect(res.status).toBe(409);
-    expect(await res.json()).toMatchObject({ error: "pick_locked" });
-  });
-
-  it("409s pick_locked exactly at the kickoff instant — the boundary is locked", async () => {
-    const { league, weekIds, gameIds, memberA } = await seedPickemLeague();
-    const weekId = weekIds.get("regular:1")!;
-    const [g1] = gameIds.get("regular:1")!;
-
-    const res = await putPicks(
-      memberA.cookie,
-      league.id,
-      weekId,
-      { picks: [{ gameId: g1, side: PICKEM_PICK_SIDE.HOME, spread: null }] },
-      appAtKickoff,
-    );
-    expect(res.status).toBe(409);
-    expect(await res.json()).toMatchObject({ error: "pick_locked" });
-  });
-
-  it("keeps a locked pick already stored when a later submission omits it, and counts it toward the cap", async () => {
-    const { league, weekIds, gameIds, memberA } = await seedPickemLeague({
-      settings: { ...DEFAULT_PICKEM_SETTINGS, picksPerWeek: 2 },
-    });
-    const weekId = weekIds.get("regular:1")!;
-    const [g1, g2, g3] = gameIds.get("regular:1")!;
-
-    // g1 kicks off first — pick it while everything is still open.
-    const pre = await putPicks(memberA.cookie, league.id, weekId, {
-      picks: [{ gameId: g1, side: PICKEM_PICK_SIDE.HOME, spread: null }],
-    });
-    expect(pre.status).toBe(200);
-
-    // g1 has since kicked off; submit a different, still-unstarted game only.
-    // The locked g1 pick must survive the replace, not be dropped for being
-    // omitted from this submission.
-    const post = await putPicks(
-      memberA.cookie,
-      league.id,
-      weekId,
-      { picks: [{ gameId: g2, side: PICKEM_PICK_SIDE.HOME, spread: null }] },
-      appAfterKickoff,
-    );
-    expect(post.status).toBe(200);
-    const postOwn = ((await post.json()) as PickemWeekPicksResponse).members.find(
+    expect(resubmit.status).toBe(200);
+    const own = ((await resubmit.json()) as PickemWeekPicksResponse).members.find(
       (m) => m.userId === memberA.user.id,
     )!;
-    expect(postOwn.picks.map((p) => p.gameId).sort()).toEqual([g1, g2].sort());
-
-    // picksPerWeek is 2: the locked g1 pick still counts toward the cap, so
-    // adding two MORE new picks (g2 again + g3) would total 3 — over the cap.
-    const overCap = await putPicks(
-      memberA.cookie,
-      league.id,
-      weekId,
-      {
-        picks: [
-          { gameId: g2, side: PICKEM_PICK_SIDE.HOME, spread: null },
-          { gameId: g3, side: PICKEM_PICK_SIDE.HOME, spread: null },
-        ],
-      },
-      appAfterKickoff,
-    );
-    expect(overCap.status).toBe(400);
-    expect(await overCap.json()).toMatchObject({ error: "too_many_picks" });
+    expect(own.picks).toHaveLength(3);
+    expect(own.picks.every((p) => p.side === PICKEM_PICK_SIDE.AWAY)).toBe(true);
   });
 
-  describe("retention keys on pickable, not locked alone", () => {
-    // The shared FOUR_GAME_WEEK's 4th, never-picked game exists purely so
-    // cancelling one of the three picked games doesn't itself shrink
-    // `picksAllowed` (min(picksPerWeek, pickable count) would otherwise drop
-    // in lockstep with `retainedCount`, tripping the cap — that interaction
-    // is covered on its own further down).
-
-    it.each([
-      {
-        label: "cancelled",
-        apply: (gameId: string) =>
-          db.update(games).set({ status: GAME_STATUS.CANCELLED }).where(eq(games.id, gameId)),
-      },
-      {
-        label: "moved",
-        apply: (gameId: string) =>
-          db.update(games).set({ overrideStatus: GAME_STATUS.MOVED }).where(eq(games.id, gameId)),
-      },
-    ])(
-      "keeps a pick on a still-unlocked $label game when a later submission omits it",
-      async ({ apply }) => {
-        const { league, weekIds, gameIds, memberA } = await seedPickemLeague({
-          settings: { ...DEFAULT_PICKEM_SETTINGS, picksPerWeek: 3 },
-          weeks: FOUR_GAME_WEEK,
-        });
-        const weekId = weekIds.get("regular:1")!;
-        const [g1, g2, g3] = gameIds.get("regular:1")!;
-
-        const initial = await putPicks(memberA.cookie, league.id, weekId, {
-          picks: [
-            { gameId: g1, side: PICKEM_PICK_SIDE.HOME, spread: null },
-            { gameId: g2, side: PICKEM_PICK_SIDE.HOME, spread: null },
-            { gameId: g3, side: PICKEM_PICK_SIDE.HOME, spread: null },
-          ],
-        });
-        expect(initial.status).toBe(200);
-
-        // g1 becomes unplayable while still unstarted — deliberately no clock
-        // advance, so `locked` stays false for g1 and only `pickable`
-        // distinguishes it. Before the fix, keying retention on `locked` alone
-        // would have deleted this pick for being omitted below.
-        await apply(g1!);
-
-        const res = await putPicks(memberA.cookie, league.id, weekId, {
-          picks: [
-            { gameId: g2, side: PICKEM_PICK_SIDE.HOME, spread: null },
-            { gameId: g3, side: PICKEM_PICK_SIDE.HOME, spread: null },
-          ],
-        });
-        expect(res.status).toBe(200);
-        const own = ((await res.json()) as PickemWeekPicksResponse).members.find(
-          (m) => m.userId === memberA.user.id,
-        )!;
-        expect(own.picks.map((p) => p.gameId).sort()).toEqual([g1, g2, g3].sort());
-      },
-    );
-
-    it("still deletes a pick on a still-scheduled, unlocked game that is simply omitted — the fix doesn't over-retain", async () => {
-      const { league, weekIds, gameIds, memberA } = await seedPickemLeague({
-        settings: { ...DEFAULT_PICKEM_SETTINGS, picksPerWeek: 3 },
-        weeks: FOUR_GAME_WEEK,
-      });
-      const weekId = weekIds.get("regular:1")!;
-      const [g1, g2, g3] = gameIds.get("regular:1")!;
-
-      const initial = await putPicks(memberA.cookie, league.id, weekId, {
-        picks: [
-          { gameId: g1, side: PICKEM_PICK_SIDE.HOME, spread: null },
-          { gameId: g2, side: PICKEM_PICK_SIDE.HOME, spread: null },
-          { gameId: g3, side: PICKEM_PICK_SIDE.HOME, spread: null },
-        ],
-      });
-      expect(initial.status).toBe(200);
-
-      await db.update(games).set({ status: GAME_STATUS.CANCELLED }).where(eq(games.id, g1!));
-
-      // Omit both the cancelled g1 (unreplaceable — must survive) and the
-      // still-scheduled g2 (an ordinary replaceable pick — must be dropped,
-      // exactly like the pre-existing wholesale-replace behavior).
-      const res = await putPicks(memberA.cookie, league.id, weekId, {
-        picks: [{ gameId: g3, side: PICKEM_PICK_SIDE.HOME, spread: null }],
-      });
-      expect(res.status).toBe(200);
-      const own = ((await res.json()) as PickemWeekPicksResponse).members.find(
-        (m) => m.userId === memberA.user.id,
-      )!;
-      expect(own.picks.map((p) => p.gameId).sort()).toEqual([g1, g3].sort());
-    });
-  });
-
-  it("never refuses an empty submission, even though a cancellation has shrunk the cap below what's retained", async () => {
+  /**
+   * The pick's game is absent from the week's slate by construction, so the
+   * client has nothing to join against to name it — every other pick renders
+   * its matchup from the slate. Without `movedGame` on the wire the only
+   * honest thing the UI could say was "a game moved", which tells a member
+   * their pick pushed but not which pick it was (manual regression, Pass 5).
+   */
+  it("carries a moved game's matchup and destination week on the pick, and null for picks still in the week", async () => {
     const { league, weekIds, gameIds, memberA } = await seedPickemLeague({
-      settings: { ...DEFAULT_PICKEM_SETTINGS, picksPerWeek: 3 },
+      weeks: [
+        ...THREE_GAME_WEEK,
+        {
+          weekNumber: 2,
+          kickoffs: [{ kickoffAt: new Date(WEEK1_KICKOFF.getTime() + 7 * 24 * 60 * 60 * 1000) }],
+        },
+      ],
     });
-    const weekId = weekIds.get("regular:1")!;
+    const week1Id = weekIds.get("regular:1")!;
+    const week2Id = weekIds.get("regular:2")!;
     const [g1, g2, g3] = gameIds.get("regular:1")!;
 
-    const initial = await putPicks(memberA.cookie, league.id, weekId, {
+    await putPicks(memberA.cookie, league.id, week1Id, {
       picks: [
         { gameId: g1, side: PICKEM_PICK_SIDE.HOME, spread: null },
         { gameId: g2, side: PICKEM_PICK_SIDE.HOME, spread: null },
         { gameId: g3, side: PICKEM_PICK_SIDE.HOME, spread: null },
       ],
     });
-    expect(initial.status).toBe(200);
+    await db.update(games).set({ weekId: week2Id }).where(eq(games.id, g1!));
 
-    // Cancelling g1 drops picksAllowed from 3 to 2 (only g2/g3 are still
-    // pickable) while the member still holds 3 rows — an empty, delete-only
-    // submission must never be refused by a cap it can only shrink against.
-    await db.update(games).set({ status: GAME_STATUS.CANCELLED }).where(eq(games.id, g1!));
+    const res = await getPicks(memberA.cookie, league.id, week1Id);
+    const own = ((await res.json()) as PickemWeekPicksResponse).members.find(
+      (m) => m.userId === memberA.user.id,
+    )!;
 
-    const res = await putPicks(memberA.cookie, league.id, weekId, { picks: [] });
-    expect(res.status).toBe(200);
-  });
+    const moved = own.picks.find((p) => p.gameId === g1)!;
+    expect(moved.movedGame).toMatchObject({ weekLabel: expect.any(String) });
+    expect(moved.movedGame?.homeTeam.abbreviation).toEqual(expect.any(String));
+    expect(moved.movedGame?.awayTeam.abbreviation).toEqual(expect.any(String));
 
-  it("bounds a submission by remaining slots (picksAllowed minus retained), not the raw picksAllowed number", async () => {
-    const { league, weekIds, gameIds, memberA } = await seedPickemLeague({
-      settings: { ...DEFAULT_PICKEM_SETTINGS, picksPerWeek: 2 },
-    });
-    const weekId = weekIds.get("regular:1")!;
-    const [g1, g2, g3] = gameIds.get("regular:1")!;
-
-    const initial = await putPicks(memberA.cookie, league.id, weekId, {
-      picks: [{ gameId: g1, side: PICKEM_PICK_SIDE.HOME, spread: null }],
-    });
-    expect(initial.status).toBe(200);
-
-    // picksAllowed stays min(2, 2 pickable) = 2 after this cancellation (g2/g3
-    // remain pickable); the retained cancelled g1 pick occupies one of those
-    // two slots, leaving exactly one remaining for a fresh submission.
-    await db.update(games).set({ status: GAME_STATUS.CANCELLED }).where(eq(games.id, g1!));
-
-    const tooMany = await putPicks(memberA.cookie, league.id, weekId, {
-      picks: [
-        { gameId: g2, side: PICKEM_PICK_SIDE.HOME, spread: null },
-        { gameId: g3, side: PICKEM_PICK_SIDE.HOME, spread: null },
-      ],
-    });
-    expect(tooMany.status).toBe(400);
-    expect(await tooMany.json()).toMatchObject({ error: "too_many_picks" });
-
-    const ok = await putPicks(memberA.cookie, league.id, weekId, {
-      picks: [{ gameId: g2, side: PICKEM_PICK_SIDE.HOME, spread: null }],
-    });
-    expect(ok.status).toBe(200);
-  });
-
-  describe("a pick whose game moved to another week", () => {
-    /** Week 1's 3-game slate plus a second week to move a game into. */
-    function twoWeekSettings(picksPerWeek: number): PickemSettings {
-      return { ...DEFAULT_PICKEM_SETTINGS, picksPerWeek };
-    }
-
-    async function seedMovableGame(picksPerWeek: number) {
-      const { league, weekIds, gameIds, memberA } = await seedPickemLeague({
-        settings: twoWeekSettings(picksPerWeek),
-        weeks: [
-          ...THREE_GAME_WEEK,
-          {
-            weekNumber: 2,
-            kickoffs: [{ kickoffAt: new Date(WEEK1_KICKOFF.getTime() + 7 * 24 * 60 * 60 * 1000) }],
-          },
-        ],
-      });
-      const week1Id = weekIds.get("regular:1")!;
-      const week2Id = weekIds.get("regular:2")!;
-      const [g1, g2, g3] = gameIds.get("regular:1")!;
-      return { league, week1Id, week2Id, g1: g1!, g2: g2!, g3: g3!, memberA };
-    }
-
-    /**
-     * The pick's game is absent from the week's slate by construction, so the
-     * client has nothing to join against to name it — every other pick renders
-     * its matchup from the slate. Without `movedGame` on the wire the only
-     * honest thing the UI could say was "a game moved", which tells a member
-     * their pick pushed but not which pick it was (manual regression, Pass 5).
-     */
-    it("carries the moved game's matchup and destination week on the pick, and null for picks still in the week", async () => {
-      const { league, week1Id, week2Id, g1, g2, memberA } = await seedMovableGame(5);
-
-      await putPicks(memberA.cookie, league.id, week1Id, {
-        picks: [
-          { gameId: g1, side: PICKEM_PICK_SIDE.HOME, spread: null },
-          { gameId: g2, side: PICKEM_PICK_SIDE.HOME, spread: null },
-        ],
-      });
-      await db.update(games).set({ weekId: week2Id }).where(eq(games.id, g1));
-
-      const res = await getPicks(memberA.cookie, league.id, week1Id);
-      const own = ((await res.json()) as PickemWeekPicksResponse).members.find(
-        (m) => m.userId === memberA.user.id,
-      )!;
-
-      const moved = own.picks.find((p) => p.gameId === g1)!;
-      expect(moved.movedGame).toMatchObject({ weekLabel: expect.any(String) });
-      expect(moved.movedGame?.homeTeam.abbreviation).toEqual(expect.any(String));
-      expect(moved.movedGame?.awayTeam.abbreviation).toEqual(expect.any(String));
-
-      // The pick that stayed put renders from the slate and must not carry it —
-      // a non-null here would make "did this move?" ambiguous.
-      expect(own.picks.find((p) => p.gameId === g2)!.movedGame).toBeNull();
-    });
-
-    it("is not deleted by a later submission that omits it", async () => {
-      const { league, week1Id, week2Id, g1, g2, memberA } = await seedMovableGame(5);
-
-      const initial = await putPicks(memberA.cookie, league.id, week1Id, {
-        picks: [
-          { gameId: g1, side: PICKEM_PICK_SIDE.HOME, spread: null },
-          { gameId: g2, side: PICKEM_PICK_SIDE.HOME, spread: null },
-        ],
-      });
-      expect(initial.status).toBe(200);
-
-      // g1 moves to week 2 — its pick keeps week 1, the week it was made in.
-      await db.update(games).set({ weekId: week2Id }).where(eq(games.id, g1));
-
-      // Submitting only g2 must not destroy the now-unaddressable g1 pick.
-      const later = await putPicks(memberA.cookie, league.id, week1Id, {
-        picks: [{ gameId: g2, side: PICKEM_PICK_SIDE.HOME, spread: null }],
-      });
-      expect(later.status).toBe(200);
-      const own = ((await later.json()) as PickemWeekPicksResponse).members.find(
-        (m) => m.userId === memberA.user.id,
-      )!;
-      expect(own.picks.map((p) => p.gameId).sort()).toEqual([g1, g2].sort());
-    });
-
-    it("survives even when the moved game has since kicked off (locked)", async () => {
-      const { league, week1Id, week2Id, g1, g2, memberA } = await seedMovableGame(5);
-
-      await putPicks(memberA.cookie, league.id, week1Id, {
-        picks: [
-          { gameId: g1, side: PICKEM_PICK_SIDE.HOME, spread: null },
-          { gameId: g2, side: PICKEM_PICK_SIDE.HOME, spread: null },
-        ],
-      });
-      await db.update(games).set({ weekId: week2Id }).where(eq(games.id, g1));
-
-      // g1's own kickoff (unchanged by the week move) is already in the past;
-      // g2's is still an hour out, so this submission is otherwise legal.
-      const later = await putPicks(
-        memberA.cookie,
-        league.id,
-        week1Id,
-        { picks: [{ gameId: g2, side: PICKEM_PICK_SIDE.HOME, spread: null }] },
-        appAfterKickoff,
-      );
-      expect(later.status).toBe(200);
-      const own = ((await later.json()) as PickemWeekPicksResponse).members.find(
-        (m) => m.userId === memberA.user.id,
-      )!;
-      expect(own.picks.map((p) => p.gameId).sort()).toEqual([g1, g2].sort());
-    });
-
-    it("still counts toward the picks-per-week cap", async () => {
-      const { league, week1Id, week2Id, g1, g2, g3, memberA } = await seedMovableGame(2);
-
-      await putPicks(memberA.cookie, league.id, week1Id, {
-        picks: [
-          { gameId: g1, side: PICKEM_PICK_SIDE.HOME, spread: null },
-          { gameId: g2, side: PICKEM_PICK_SIDE.HOME, spread: null },
-        ],
-      });
-      await db.update(games).set({ weekId: week2Id }).where(eq(games.id, g1));
-
-      // The retained (moved-out) g1 pick occupies one of the two slots, so
-      // two brand-new picks push the total to three — over the cap.
-      const res = await putPicks(memberA.cookie, league.id, week1Id, {
-        picks: [
-          { gameId: g2, side: PICKEM_PICK_SIDE.HOME, spread: null },
-          { gameId: g3, side: PICKEM_PICK_SIDE.HOME, spread: null },
-        ],
-      });
-      expect(res.status).toBe(400);
-      expect(await res.json()).toMatchObject({ error: "too_many_picks" });
-    });
+    // The pick that stayed put renders from the slate and must not carry it —
+    // a non-null here would make "did this move?" ambiguous.
+    expect(own.picks.find((p) => p.gameId === g2)!.movedGame).toBeNull();
   });
 
   it("rolls back the whole submission when one entry is refused — no partial write", async () => {
@@ -1025,38 +863,36 @@ describe("PUT /api/leagues/:leagueId/pickem/weeks/:weekId/picks", () => {
       ],
     });
     const week1Id = weekIds.get("regular:1")!;
-    const [g1, g2, g3] = gameIds.get("regular:1")!;
+    const [g1, g2] = gameIds.get("regular:1")!;
     const week2GameId = gameIds.get("regular:2")![0]!;
 
-    const initial = await putPicks(memberA.cookie, league.id, week1Id, {
+    // The right size for week 1, but the third entry references a game from a
+    // different week — the whole submission must be refused, and since a
+    // member gets only one, a partial write would cost them the week.
+    const res = await putPicks(memberA.cookie, league.id, week1Id, {
       picks: [
         { gameId: g1, side: PICKEM_PICK_SIDE.HOME, spread: null },
         { gameId: g2, side: PICKEM_PICK_SIDE.HOME, spread: null },
-      ],
-    });
-    expect(initial.status).toBe(200);
-
-    // The third entry references a game from a different week — the whole
-    // submission must be refused, not just that one entry.
-    const res = await putPicks(memberA.cookie, league.id, week1Id, {
-      picks: [
-        { gameId: g1, side: PICKEM_PICK_SIDE.AWAY, spread: null },
-        { gameId: g3, side: PICKEM_PICK_SIDE.HOME, spread: null },
         { gameId: week2GameId, side: PICKEM_PICK_SIDE.HOME, spread: null },
       ],
     });
     expect(res.status).toBe(400);
     expect(await res.json()).toMatchObject({ error: "game_not_in_week" });
 
-    const after = await getPicks(memberA.cookie, league.id, week1Id);
-    const own = ((await after.json()) as PickemWeekPicksResponse).members.find(
-      (m) => m.userId === memberA.user.id,
-    )!;
-    expect(own.picks.map((p) => p.gameId).sort()).toEqual([g1, g2].sort());
-    expect(own.picks.find((p) => p.gameId === g1)?.side).toBe(PICKEM_PICK_SIDE.HOME);
+    expect(await pickCountFor(league.id)).toBe(0);
+
+    // And the week is still open: nothing landed, so nothing was submitted.
+    const retry = await putPicks(memberA.cookie, league.id, week1Id, {
+      picks: gameIds.get("regular:1")!.map((gameId) => ({
+        gameId,
+        side: PICKEM_PICK_SIDE.HOME,
+        spread: null,
+      })),
+    });
+    expect(retry.status).toBe(200);
   });
 
-  it("400s too_many_picks when submitting more than picksPerWeek", async () => {
+  it("400s too_many_picks when the set is larger than the week's required size", async () => {
     const { league, weekIds, gameIds, memberA } = await seedPickemLeague({
       settings: { ...DEFAULT_PICKEM_SETTINGS, picksPerWeek: 2 },
     });
@@ -1074,25 +910,17 @@ describe("PUT /api/leagues/:leagueId/pickem/weeks/:weekId/picks", () => {
     expect(await res.json()).toMatchObject({ error: "too_many_picks" });
   });
 
-  // NOTE: a genuine "slate shorter than picksPerWeek AND *this submission's own
-  // games* exceed the slate size" case is unreachable through the API as
-  // currently written — any submission whose games are all valid, distinct,
-  // and unlocked is bounded above by the slate's own game count, so it can
-  // never exceed picksAllowed (== slate.length in that regime) without first
-  // tripping game_not_in_week or duplicate_pick, both of which are checked
-  // ahead of the cap. The cap IS reachable in a short slate, though — via
-  // *retained* picks (locked, above, or moved-out, below) occupying slots the
-  // submission itself doesn't ask for.
-
   it("400s duplicate_pick when the same gameId appears twice", async () => {
     const { league, weekIds, gameIds, memberA } = await seedPickemLeague();
     const weekId = weekIds.get("regular:1")!;
-    const [g1] = gameIds.get("regular:1")!;
+    const [g1, g2] = gameIds.get("regular:1")!;
 
+    // The right number of entries for the week, but only two distinct games.
     const res = await putPicks(memberA.cookie, league.id, weekId, {
       picks: [
         { gameId: g1, side: PICKEM_PICK_SIDE.HOME, spread: null },
         { gameId: g1, side: PICKEM_PICK_SIDE.AWAY, spread: null },
+        { gameId: g2, side: PICKEM_PICK_SIDE.HOME, spread: null },
       ],
     });
     expect(res.status).toBe(400);
@@ -1101,10 +929,12 @@ describe("PUT /api/leagues/:leagueId/pickem/weeks/:weekId/picks", () => {
 
   it("400s game_not_in_week when a submitted game belongs to a different week", async () => {
     // Both weeks share one season instance so this exercises game_not_in_week
-    // specifically, not week_out_of_range (a season mismatch).
+    // specifically, not week_out_of_range (a season mismatch). Week 1 holds a
+    // single game, so one entry is the full required set and the refusal can't
+    // be confused with a size complaint.
     const { league, weekIds, gameIds, memberA } = await seedPickemLeague({
       weeks: [
-        ...THREE_GAME_WEEK,
+        { weekNumber: 1, kickoffs: [{ kickoffAt: WEEK1_KICKOFF }] },
         {
           weekNumber: 2,
           kickoffs: [{ kickoffAt: new Date(WEEK1_KICKOFF.getTime() + 7 * 24 * 60 * 60 * 1000) }],
@@ -1126,10 +956,13 @@ describe("PUT /api/leagues/:leagueId/pickem/weeks/:weekId/picks", () => {
       status: LEAGUE_STATUS.CONCLUDED,
     });
     const weekId = weekIds.get("regular:1")!;
-    const [g1] = gameIds.get("regular:1")!;
 
     const res = await putPicks(memberA.cookie, league.id, weekId, {
-      picks: [{ gameId: g1, side: PICKEM_PICK_SIDE.HOME, spread: null }],
+      picks: gameIds.get("regular:1")!.map((gameId) => ({
+        gameId,
+        side: PICKEM_PICK_SIDE.HOME,
+        spread: null,
+      })),
     });
     expect(res.status).toBe(409);
     expect(await res.json()).toMatchObject({ error: "league_concluded" });
@@ -1200,7 +1033,7 @@ describe("PUT /api/leagues/:leagueId/pickem/weeks/:weekId/picks", () => {
       expect(await res.json()).toMatchObject({ error: "spread_stale" });
     });
 
-    it("409s spread_unavailable when the game has no odds snapshot at all", async () => {
+    it("409s spread_unavailable when the game has no spread at all", async () => {
       const { seasonId, weekIds, gameIds } = await seedSeason(db, {
         year: 2026,
         weeks: [{ weekNumber: 1, kickoffs: [{ kickoffAt: WEEK1_KICKOFF }] }],
@@ -1227,15 +1060,19 @@ describe("PUT /api/leagues/:leagueId/pickem/weeks/:weekId/picks", () => {
   it("ignores and nulls out a submitted spread in a straight-up league", async () => {
     const { league, weekIds, gameIds, memberA } = await seedPickemLeague();
     const weekId = weekIds.get("regular:1")!;
-    const [g1] = gameIds.get("regular:1")!;
+    const [g1, g2, g3] = gameIds.get("regular:1")!;
 
     const res = await putPicks(memberA.cookie, league.id, weekId, {
-      picks: [{ gameId: g1, side: PICKEM_PICK_SIDE.HOME, spread: -7 }],
+      picks: [
+        { gameId: g1, side: PICKEM_PICK_SIDE.HOME, spread: -7 },
+        { gameId: g2, side: PICKEM_PICK_SIDE.HOME, spread: null },
+        { gameId: g3, side: PICKEM_PICK_SIDE.HOME, spread: null },
+      ],
     });
     expect(res.status).toBe(200);
     const body = (await res.json()) as PickemWeekPicksResponse;
     const own = body.members.find((m) => m.userId === memberA.user.id)!;
-    expect(own.picks[0]).toMatchObject({ gameId: g1, spread: null });
+    expect(own.picks.find((p) => p.gameId === g1)).toMatchObject({ spread: null });
   });
 });
 
@@ -1244,12 +1081,16 @@ describe("PATCH /api/leagues/:leagueId — settings changes reset picks (setting
     return { ...DEFAULT_PICKEM_SETTINGS, ...overrides };
   }
 
-  async function seedWithOnePick() {
+  /** One member, one submitted week — the week's full three-game set. */
+  async function seedWithSubmittedWeek() {
     const { league, weekIds, gameIds, memberA } = await seedPickemLeague();
     const weekId = weekIds.get("regular:1")!;
-    const [g1] = gameIds.get("regular:1")!;
     const res = await putPicks(memberA.cookie, league.id, weekId, {
-      picks: [{ gameId: g1, side: PICKEM_PICK_SIDE.HOME, spread: null }],
+      picks: gameIds.get("regular:1")!.map((gameId) => ({
+        gameId,
+        side: PICKEM_PICK_SIDE.HOME,
+        spread: null,
+      })),
     });
     expect(res.status).toBe(200);
     return { league, memberA };
@@ -1274,7 +1115,7 @@ describe("PATCH /api/leagues/:leagueId — settings changes reset picks (setting
       settings: settingsWith({ endWeek: { type: WEEK_TYPE.REGULAR, number: 10 } }),
     },
   ])("clears picks when $label — a change that could strand them", async ({ settings }) => {
-    const { league, memberA } = await seedWithOnePick();
+    const { league, memberA } = await seedWithSubmittedWeek();
 
     const res = await patchLeague(memberA.cookie, league.id, { settings });
     expect(res.status).toBe(200);
@@ -1287,11 +1128,11 @@ describe("PATCH /api/leagues/:leagueId — settings changes reset picks (setting
       settings: settingsWith({ endWeek: { type: WEEK_TYPE.POSTSEASON, number: 4 } }),
     },
   ])("keeps picks when $label — nothing is stranded", async ({ settings }) => {
-    const { league, memberA } = await seedWithOnePick();
+    const { league, memberA } = await seedWithSubmittedWeek();
 
     const res = await patchLeague(memberA.cookie, league.id, { settings });
     expect(res.status).toBe(200);
-    expect(await pickCountFor(league.id)).toBe(1);
+    expect(await pickCountFor(league.id)).toBe(3);
   });
 
   it("clears picks when a stored settings row omits picksPerWeek entirely and the new value undercuts the schema default", async () => {
